university-cms/backend/requirements.txt
Flask
Flask-CORS
python-dotenv
bcrypt
PyJWT
